+++ packages/Angular/Explorer/dashboards/src/Integration/components/pipelines/pipelines.component.ts
@@ -165,6 +165,7 @@ export class PipelinesComponent extends BaseResourceComponent implements OnInit,
 
   async ngOnInit(): Promise<void> {
     await this.LoadData();
+    this.NotifyLoadComplete();
   }
 
   ngOnDestroy(): void {