+++ packages/AI/Vectors/Providers/Pinecone/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log - @memberjunction/ai-vectors-pinecone
 
+## 5.24.0
+
+### Minor Changes
+
+- c318a0c: metadata + migrations in this PR == minor
+
+### Patch Changes
+
+- Updated dependencies [c318a0c]
+- Updated dependencies [1912726]
+  - @memberjunction/ai-vectors@5.24.0
+  - @memberjunction/ai-vectordb@5.24.0
+  - @memberjunction/core@5.24.0
+  - @memberjunction/aiengine@5.24.0
+  - @memberjunction/global@5.24.0
+
 ## 5.23.0
 
 ### Minor Changes