+++ packages/Angular/Generic/chat/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log - @memberjunction/ng-chat
 
+## 5.24.0
+
+### Patch Changes
+
+- Updated dependencies [c318a0c]
+- Updated dependencies [1912726]
+  - @memberjunction/core@5.24.0
+  - @memberjunction/ng-container-directives@5.24.0
+  - @memberjunction/ng-shared-generic@5.24.0
+  - @memberjunction/ng-markdown@5.24.0
+  - @memberjunction/ng-ui-components@5.24.0
+
 ## 5.23.0
 
 ### Patch Changes