+++ packages/Angular/Generic/container-directives/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log - @memberjunction/ng-container-directives
 
+## 5.24.0
+
+### Patch Changes
+
+- Updated dependencies [c318a0c]
+- Updated dependencies [1912726]
+  - @memberjunction/core@5.24.0
+  - @memberjunction/global@5.24.0
+
 ## 5.23.0
 
 ### Patch Changes