+++ packages/Angular/Generic/entity-relationship-diagram/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @memberjunction/ng-entity-relationship-diagram
 
+## 5.24.0
+
+### Patch Changes
+
+- Updated dependencies [c318a0c]
+- Updated dependencies [1912726]
+  - @memberjunction/core@5.24.0
+  - @memberjunction/ng-ui-components@5.24.0
+
 ## 5.23.0
 
 ### Patch Changes