+++ packages/Angular/Generic/gantt/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @memberjunction/ng-gantt
 
+## 5.24.0
+
+### Patch Changes
+
+- @memberjunction/global@5.24.0
+
 ## 5.23.0
 
 ### Patch Changes