+++ packages/MJGlobal/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log - @memberjunction/global
 
+## 5.24.0
+
 ## 5.23.0
 
 ### Patch Changes