+++ packages/MJInstaller/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @memberjunction/installer
 
+## 5.24.0
+
 ## 5.23.0
 
 ### Patch Changes