+++ packages/SQLServerDataProvider/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log - @memberjunction/sqlserver-dataprovider
 
+## 5.24.0
+
+### Patch Changes
+
+- Updated dependencies [c318a0c]
+- Updated dependencies [1912726]
+  - @memberjunction/ai-vector-dupe@5.24.0
+  - @memberjunction/core@5.24.0
+  - @memberjunction/core-entities@5.24.0
+  - @memberjunction/aiengine@5.24.0
+  - @memberjunction/actions@5.24.0
+  - @memberjunction/ai-provider-bundle@5.24.0
+  - @memberjunction/actions-base@5.24.0
+  - @memberjunction/encryption@5.24.0
+  - @memberjunction/generic-database-provider@5.24.0
+  - @memberjunction/queue@5.24.0
+  - @memberjunction/query-processor@5.24.0
+  - @memberjunction/ai@5.24.0
+  - @memberjunction/global@5.24.0
+  - @memberjunction/sql-dialect@5.24.0
+
 ## 5.23.0
 
 ### Patch Changes