+++ packages/Templates/engine/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log - @memberjunction/templates
 
+## 5.24.0
+
+### Patch Changes
+
+- Updated dependencies [c318a0c]
+- Updated dependencies [1912726]
+  - @memberjunction/ai-core-plus@5.24.0
+  - @memberjunction/core@5.24.0
+  - @memberjunction/core-entities@5.24.0
+  - @memberjunction/aiengine@5.24.0
+  - @memberjunction/ai-provider-bundle@5.24.0
+  - @memberjunction/templates-base-types@5.24.0
+  - @memberjunction/ai@5.24.0
+  - @memberjunction/global@5.24.0
+
 ## 5.23.0
 
 ### Patch Changes